Updated Logistic Regression and PCA, not RF
Logistic Regression and PCA should be complete now. I left the old PCA there just in case there was a preference for that, but I think it can be deleted. Words might need to be edited!
Could not figure out random forest errors. spent a decent amount of time on it, but kept getting stuck. really sorry.